requirements.txt
sentence-transformers
xformers
chromadb
pandas
numpy
matplotlib

